refactored scheduler and monitor.

+++ sys/daemonize.h
@@ -25,10 +25,10 @@
    Handling process creation.
 */
 void single_instance_daemonize(const char *lockfolder_in, 
-			       const char *progname,long version,
+			       const char *progname, int version,
 			       void(*daemon_func)(void*), 
 			       void* daemon_arg);
-int lock_file(const char *fn, long block, long version);
+int lock_file(const char *fn, int block, int version);
 void daemonize(void);
 void redirect(void);
 pid_t launch_exe(const char *exepath, const char *cmd);
